app/requirements.txt
pyspark==3.5.0
boto3==1.34.0
opensearch-py==2.4.2
redis==5.0.1
jellyfish==1.0.3
python-Levenshtein==0.23.0
phonetics==1.0.5
recordlinkage==0.16
pandas==2.1.4
pyarrow==14.0.1
awswrangler==3.5.2
